fix: Strip out ParallelBranchTracker and fold into Base-state
