+++ lib/index.js
@@ -12,7 +12,6 @@ const resources = require('./resources')
 const MemoryDao = require('./dao/Memory-dao')
 const StorageDao = require('./dao/StorageService-dao')
 const Status = require('./Status')
-const ParallelBranchTracker = require('./Parallel-branch-tracker')
 const CallbackManager = require('./Callback-manager')
 const debug = require('debug')('statebox')
 
@@ -25,7 +24,6 @@ class Statebox {
         this.options.dao = dao
         this.options.executioner = executioner
         this.options.callbackManager = new CallbackManager()
-        this.options.parallelBranchTracker = new ParallelBranchTracker(dao)
       })
   }
