+++ lib/state-machines/state-types/Base-state.js
@@ -14,7 +14,6 @@ class BaseState {
     this.stateMachineName = stateMachine.name
     this.definition = stateDefinition
     this.dao = options.dao
-    this.parallelBranchTracker = options.parallelBranchTracker
     this.callbackManager = options.callbackManager
 
     this.inputSelector = findSelector(stateDefinition.InputPath)
@@ -81,22 +80,24 @@ class BaseState {
       errorInfo
     )
 
-    const parentExecutionName = executionDescription.executionOptions.parentExecutionName
-    if (parentExecutionName) {
-      // Has a parent flow, so see if the related parallel state can advance
-      const tracker = this.parallelBranchTracker
-      tracker.registerChildExecutionFail(executionName)
-      this.processTaskFailure(
-        {
-          error: 'States.BranchFailed',
-          cause: errorInfo.cause || 'Failed because a state in a parallel branch has failed'
-        },
-        parentExecutionName
-      )
-    }
+    this.handleChildExecutionFail(executionDescription, errorInfo)
 
     return failedExecutionDescription
   } // runTaskFailure
+
+  handleChildExecutionFail (executionDescription, errorInfo) {
+    const parentExecutionName = executionDescription.executionOptions.parentExecutionName
+    if (!parentExecutionName) return
+
+    debugPackage(`Branch ${executionDescription.executionName} failed in  ${parentExecutionName}`)
+    this.processTaskFailure(
+      {
+        error: 'States.BranchFailed',
+        cause: errorInfo.cause || 'Failed because a state in a parallel branch has failed'
+      },
+      parentExecutionName
+    )
+  } // handleChildExecutionFail
 
   hasCatchHandler (stateDefinition, errorRaised) {
     if (!stateDefinition.Catch) {
@@ -128,8 +129,6 @@ class BaseState {
 
   async runTaskHeartbeat (executionDescription, output) {
     const executionName = executionDescription.executionName
-    const tracker = this.parallelBranchTracker
-    tracker.registerChildExecutionFail(executionName)
     executionDescription.ctx = _.defaults(output, executionDescription.ctx)
 
     await this.dao.setNextState(
@@ -158,7 +157,7 @@ class BaseState {
     const stateDefinition = stateMachine.findStateDefinitionByName(executionDescription.currentStateName)
 
     // END
-    if (stateDefinition.End || stateDefinition.Type === 'Succeed') {
+    if (stateDefinition.End || stateDefinition.Type === 'Succeed' || stateDefinition.Type === 'Fail') {
       return this.runTaskEnd(executionName, executionDescription, output)
     } else {
       // NEXT
@@ -171,29 +170,45 @@ class BaseState {
       executionName,
       ctx
     )
-    const parentExecutionName = executionDescription.executionOptions.parentExecutionName
-    if (parentExecutionName) {
-      // Has a parent flow, so see if the related parallel state can advance
-      const tracker = this.parallelBranchTracker
-      const [parentCompleted, parentCtx] = await tracker.registerChildExecutionEnd(executionDescription, ctx)
-      if (parentCompleted) {
-        debugPackage(`All branches have now succeeded (executionName='${executionDescription.executionOptions.parentExecutionName}')`)
-        this.processTaskSuccess(parentExecutionName, parentCtx)
-      }
-    } else {
-      // No branching, so finished everything... might need to call a callback?
-      this.callbackManager.fireCallback(
-        Status.COMPLETE,
-        executionName,
-        succeededExecutionDescription
-      )
-    }
+
+    this.handleChildExecutionSuccess(executionDescription)
+
+    this.callbackManager.fireCallback(
+      Status.COMPLETE,
+      executionName,
+      succeededExecutionDescription
+    )
 
     return succeededExecutionDescription
   } // runTaskEnd
 
+  async handleChildExecutionSuccess (executionDescription) {
+    const parentExecutionName = executionDescription.executionOptions.parentExecutionName
+
+    if (!parentExecutionName) return
+
+    const allChildren = await this.dao.findExecutionsByParentName(parentExecutionName)
+    const completedCount = allChildren.filter(child => child.status !== Status.RUNNING).length
+    const failedCount = allChildren.filter(child => child.status === Status.FAILED).length
+
+    const parentCompleted = (completedCount === allChildren.length)
+    const parentSucceeded = parentCompleted && (failedCount === 0)
+
+    if (parentCompleted) {
+      debugPackage(`All branches have now succeeded in ${parentExecutionName}`)
+
+      const parentCtx = makeParentContext(allChildren)
+      this.processTaskSuccess(parentExecutionName, parentCtx)
+    } else if (parentSucceeded) {
+      debugPackage(`All branches have now completed in ${parentExecutionName} with ${failedCount} failures`)
+    } else {
+      debugPackage(`${completedCount} of ${allChildren.length} completed in ${parentExecutionName}`)
+    }
+  } // handleChildExecutionSuccess
+
   async runTaskNextState (executionName, stateMachine, nextState, ctx) {
-    const nextResource = stateMachine.findStateDefinitionByName(nextState).Resource
+    const nextDefinition = stateMachine.findStateDefinitionByName(nextState)
+    const nextResource = nextDefinition.Resource
     await this.dao.setNextState(
       executionName, // executionName
       nextState, // nextStateName
@@ -266,5 +281,16 @@ function applyResultPath (rp, ctx, result) {
   }
   return ctx
 } // applyResultPath
+
+function makeParentContext (allChildren) {
+  const ctx = [ ]
+  for (const child of allChildren) {
+    const branchIndex = child.executionOptions.branchIndex
+    const childCtx = child.ctx
+
+    ctx[branchIndex] = childCtx
+  }
+  return ctx
+}
 
 module.exports = BaseState
