+++ lib/state-machines/state-types/Parallel.js
@@ -22,7 +22,8 @@ class Parallel extends BaseStateType {
   async process (executionDescription, input) {
     const parentExecutionName = executionDescription.executionName
 
-    await this.parallelBranchTracker.addParentExecution(executionDescription, this.branches.length)
+    executionDescription.childCount = this.branches.length
+    await this.dao.checkpoint(executionDescription)
 
     const branchExecutions = this.branches
       .map((stateMachineName, index) => {
